+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+2.0.0
+: Rely on Redcarpet for Markdown parsing, far more accurate with a few losses I'll handle over time
+: Config file at ~/.config/mdless/config.yml
+: Allow inlining of footnotes
+: Nested list indentation
+
 1.0.37
 : Comments inside of fenced code rendering as ATX headlines